Add webextension for asynchronous backchannel from browser.
This does not work yet. For some reason, content.js does not load when
the extension is installed via marionette. If it is installed manually
as temporary extension in `about:debugging`, everything works. If you
install it via marionette, it does not run and you see errors like:

Content Security Policy: The page’s settings blocked the loading of a
resource at inline (“default-src”).

in the browser console. I could not figure out yet why this happens and
how to resolve it. One idea is to not use an extension after all but to
inject the javascript code via marionette manually.